+++ Discord_Bot/queries.py
